rename FilteredEventParameter to Filter and move it into model
