+++ restservice/src/main/java/org/itevents/dao/EventDao.java
@@ -1,7 +1,7 @@
 package org.itevents.dao;
 
 import org.itevents.model.Event;
-import org.itevents.parameter.FilteredEventsParameter;
+import org.itevents.model.Filter;
 
 import java.util.List;
 
@@ -17,7 +17,7 @@ public interface EventDao {
 
     void updateEvent(Event event);
 
-    List<Event> getFilteredEvents(FilteredEventsParameter params);
+    List<Event> getFilteredEvents(Filter params);
     
     void removeEvent(Event event);
