+++ restservice/src/main/java/org/itevents/dao/mybatis/util/GetFilteredEventsSqlBuilder.java
@@ -1,15 +1,15 @@
 package org.itevents.dao.mybatis.util;
 
 import org.apache.ibatis.jdbc.SQL;
+import org.itevents.model.Filter;
 import org.itevents.model.Technology;
-import org.itevents.parameter.FilteredEventsParameter;
 
 import java.util.Iterator;
 import java.util.List;
 
 public class GetFilteredEventsSqlBuilder {
 
-    public String getFilteredEvents(final FilteredEventsParameter params) {
+    public String getFilteredEvents(final Filter params) {
         return new SQL() {{
             SELECT("*");
             FROM("event e");
@@ -34,7 +34,7 @@ public class GetFilteredEventsSqlBuilder {
         }}.toString() + " ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
     }
 
-    private String makeJoin(FilteredEventsParameter params) {
+    private String makeJoin(Filter params) {
         StringBuilder sb = new StringBuilder();
         sb.append("event_technology et ON ");
         List<Technology> technologies = params.getTechnologies();
