+++ restservice/src/test/java/org/itevents/dao/mybatis/mapper/EventMapperDbTest.java
@@ -7,8 +7,8 @@ import com.github.springtestdbunit.annotation.ExpectedDatabase;
 import com.github.springtestdbunit.assertion.DatabaseAssertionMode;
 import org.itevents.AbstractDbTest;
 import org.itevents.model.Event;
+import org.itevents.model.Filter;
 import org.itevents.model.Technology;
-import org.itevents.parameter.FilteredEventsParameter;
 import org.itevents.service.converter.EventConverter;
 import org.itevents.test_utils.BuilderUtil;
 import org.itevents.wrapper.EventWrapper;
@@ -98,14 +98,14 @@ public class EventMapperDbTest extends AbstractDbTest {
 
     @Test
     public void shouldFindEventsWithDefaultParameters() {
-        FilteredEventsParameter parameter = getDefaultFilteredEventsParameter();
+        Filter parameter = getDefaultFilteredEventsParameter();
         int expectedSize = 7;
         int returnedSize = eventMapper.getFilteredEvents(parameter).size();
         assertEquals(expectedSize, returnedSize);
     }
 
-    private FilteredEventsParameter getDefaultFilteredEventsParameter() {
-        FilteredEventsParameter result = new FilteredEventsParameter();
+    private Filter getDefaultFilteredEventsParameter() {
+        Filter result = new Filter();
         result.setLimit(10);
         result.setOffset(0);
         return result;
@@ -116,7 +116,7 @@ public class EventMapperDbTest extends AbstractDbTest {
         EventWrapper wrapper = new EventWrapper();
         wrapper.setPage(3);
         wrapper.setItemsPerPage(2);
-        FilteredEventsParameter parameter = new EventConverter().convert(wrapper);
+        Filter parameter = new EventConverter().convert(wrapper);
 
         List<Event> expectedEvents = new ArrayList<>();
         expectedEvents.add(BuilderUtil.buildEventCplus());
@@ -131,7 +131,7 @@ public class EventMapperDbTest extends AbstractDbTest {
         EventWrapper wrapper = new EventWrapper();
         wrapper.setPage(30);
         wrapper.setItemsPerPage(-2);
-        FilteredEventsParameter parameter = new EventConverter().convert(wrapper);
+        Filter parameter = new EventConverter().convert(wrapper);
 
         List<Event> expectedEvents = new ArrayList<>();
 
@@ -141,7 +141,7 @@ public class EventMapperDbTest extends AbstractDbTest {
 
     @Test
     public void shouldFindEventsInKyivWithTechnologyJava() throws ParseException {
-        FilteredEventsParameter parameter = getDefaultFilteredEventsParameter();
+        Filter parameter = getDefaultFilteredEventsParameter();
         List<Technology> technologies = new ArrayList<>();
         technologies.add(BuilderUtil.buildTechnologyJava());
         parameter.setTechnologies(technologies);
@@ -156,7 +156,7 @@ public class EventMapperDbTest extends AbstractDbTest {
 
     @Test
     public void shouldFindPayedEventsInBoyarka() throws ParseException {
-        FilteredEventsParameter parameter = getDefaultFilteredEventsParameter();
+        Filter parameter = getDefaultFilteredEventsParameter();
         parameter.setCity(BuilderUtil.buildCityBoyarka());
         parameter.setFree(false);
 
@@ -173,7 +173,7 @@ public class EventMapperDbTest extends AbstractDbTest {
         technologies.add(BuilderUtil.buildTechnologyPhp());
         technologies.add(BuilderUtil.buildTechnologyAnt());
         technologies.add(BuilderUtil.buildTechnologySql());
-        FilteredEventsParameter parameter = getDefaultFilteredEventsParameter();
+        Filter parameter = getDefaultFilteredEventsParameter();
         parameter.setTechnologies(technologies);
 
         List<Event> expectedEvents = new ArrayList<>();
@@ -195,7 +195,7 @@ public class EventMapperDbTest extends AbstractDbTest {
         expectedEvents.add(BuilderUtil.buildEventPhp());
         expectedEvents.add(BuilderUtil.buildEventJs());
 
-        FilteredEventsParameter parameter = getDefaultFilteredEventsParameter();
+        Filter parameter = getDefaultFilteredEventsParameter();
         parameter.setLatitude(testLatitude);
         parameter.setLongitude(testLongitude);
         parameter.setRadius(testRadius);
