+++ restservice/src/test/java/org/itevents/service/transactional/MyBatisEventServiceTest.java
@@ -2,7 +2,7 @@ package org.itevents.service.transactional;
 
 import org.itevents.dao.EventDao;
 import org.itevents.model.Event;
-import org.itevents.parameter.FilteredEventsParameter;
+import org.itevents.model.Filter;
 import org.itevents.service.EventService;
 import org.itevents.test_utils.BuilderUtil;
 import org.itevents.wrapper.EventWrapper;
@@ -93,7 +93,7 @@ public class MyBatisEventServiceTest {
     public void shouldFindEventsByParameter() throws ParseException {
         List<Event> expectedEvents = new ArrayList<>();
         expectedEvents.add(BuilderUtil.buildEventJava());
-        when(eventDao.getFilteredEvents(any(FilteredEventsParameter.class))).thenReturn(expectedEvents);
+        when(eventDao.getFilteredEvents(any(Filter.class))).thenReturn(expectedEvents);
         EventWrapper wrapper = new EventWrapper();
         List<Event> returnedEvents = eventService.getFilteredEvents(wrapper);
         assertEquals(expectedEvents, returnedEvents);
@@ -102,7 +102,7 @@ public class MyBatisEventServiceTest {
     @Test
     public void shouldNotFindEventsByParameter() throws ParseException {
         List<Event> expectedEvents = new ArrayList<>();
-        when(eventDao.getFilteredEvents(any(FilteredEventsParameter.class))).thenThrow(Exception.class);
+        when(eventDao.getFilteredEvents(any(Filter.class))).thenThrow(Exception.class);
         EventWrapper wrapper = new EventWrapper();
         List<Event> returnedEvents = eventService.getFilteredEvents(wrapper);
         assertEquals(expectedEvents, returnedEvents);
